+++ src/app/usuarios/page.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function users({ uid }) {
+  return (
+    <div>
+      <h1>{uid}</h1>
+    </div>
+  );
+}
